Merge from Cleveland Clinic Mercurial maintenance branch.
Original metadata, including description:

changeset:   8d6eca463a31
user:        Chimezie Ogbuji <ogbujic@ccf.org>
date:        Mon May 05 16:17:32 2008 -0400
description:
Proxy commit for John clark:

Significant refactoring for modularity

This refactoring targets the FOPL relational model, and is intended to allow
for additional SQL backends to implement this model. As part of this
effort, this commit includes a PostgreSQL implementation of this model.
This refactoring also allows for better control over the management of
indices and foreign key references on the relational tables that implement
the model.


git-svn-id: http://rdflib.googlecode.com/svn/trunk@1629 4f84a9ca-5c3e-0410-8a68-0bdf370325d5